125 피격 시각 피드백 구현

+++ Assets/Scripts/Common/Interface/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int dmg);
+    void TakeDamage(float dmg);
 }
